+++ CHANGES.rst
@@ -3,6 +3,14 @@ CHANGES
 =======
 
 
+Future version
+--------------
+
+*Release date: future*
+
+- Optimized the ``copy`` method of frozen (immutable) dictionaries: without update arguments ``copy`` returns ``self``.
+
+
 v0.0.0-beta
 -----------